chore: unify emission summary schemas and update related components
